requirements.txt
asgiref==3.7.2
asttokens==2.4.1
decorator==5.1.1
Django==3.2.7
djangorestframework==3.12.4
exceptiongroup==1.2.0
executing==2.0.1
Faker==22.6.0
gunicorn==19.9.0
ipdb==0.13.13
ipython==8.21.0
jedi==0.19.1
matplotlib-inline==0.1.6
mysqlclient==2.2.4
parso==0.8.3
pexpect==4.9.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
python-dateutil==2.8.2
python-decouple==3.6
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
tomli==2.0.1
traitlets==5.14.1
typing_extensions==4.9.0
wcwidth==0.2.13
